fix : inputview에서 해야할 일이 controller에서 구현 수정
- input에 대한 검증과, 데이터 가공을 inputview에서 구현
